refactor : 매직넘버 처리, forEach 사용 (output)
