+++ src/main/java/lotto/view/output/Output.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.view.output;
 
 public class Output {
     public static final Integer LOTTO_PRICE = 1000;
