+++ src/main/java/lotto/view/output/OutputRate.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.view.output;
 
 public class OutputRate {
